Add Batch and modification methods
Batch is used to collect modifications so they can be committed.
Now that batch is implemented, we can use it for helper methods such as create, set,
merge, update, etc.
Allow Document::Reference objects that are created by the Batch object to
use the Batch when calling modification methods on the doc ref object.